Get rid of some warnings from the 64 bit compile and tidy up zargon_functions.h (formerly select_asm_or_c_function.h or something like that). Sadly running the full test suite is broken, something went wrong as we made progressive enhancements that we didn't notice since the 'c -3' test does work. We'll have to use git and bisect to find the failure point. The weird thing with the uint16_t pin_count working and the uint8_t pin_count not working could be related.

+++ src/sargon-minimax.cpp
@@ -554,7 +554,7 @@ private:
                     size_t next = offset + key.length();
                     if( next < s.length() && 'A'<= s[next] && s[next]<='H' ) 
                         continue;   // eg key = "AG" found "AGH", keep looking
-                    int prev = offset - 1; // int in case offset = 0
+                    int prev = (int)(offset - 1); // int in case offset = 0
                     if( prev >= 0 && 'A'<= s[prev] && s[prev]<='H' ) 
                         continue;   // eg key = "B" found "AGB", keep looking
                     idx = i;
